Split the `ty` tests by the property they check
