+++ tests/result/ty/ArithTinyTest
@@ -1,0 +1,1 @@
+ty.ArithTinyTest: 4 / 4
