+++ tests/result/ty/ArithTinyTest.json
@@ -1,0 +1,6 @@
+{
+  "arithmetic laws" : true,
+  "exponentiation" : true,
+  "modulo" : true,
+  "number arithmetic laws" : true
+}
