--- tests/result/ty/LatticeTinyTest
@@ -1,1 +1,0 @@
-ty.LatticeTinyTest: 9 / 9
