--- tests/result/ty/LatticeTinyTest.json
@@ -1,11 +1,0 @@
-{
-  "ast lattice laws" : true,
-  "bool lattice laws" : true,
-  "bounded set lattice laws" : true,
-  "closure lattice laws" : true,
-  "every component of `Any` is top" : true,
-  "list lattice laws" : true,
-  "map lattice laws" : true,
-  "record lattice laws" : true,
-  "value lattice laws" : true
-}
