+++ tests/result/ty/LawTinyTest
@@ -1,0 +1,1 @@
+ty.LawTinyTest: 23 / 23
