+++ tests/result/ty/LawTinyTest.json
@@ -1,0 +1,25 @@
+{
+  "NaN" : true,
+  "and" : true,
+  "ast lattice laws" : true,
+  "bool lattice laws" : true,
+  "bounded set lattice laws" : true,
+  "canonicalize" : true,
+  "closure lattice laws" : true,
+  "every component of `Any` is top" : true,
+  "int lattice laws" : true,
+  "join across representations" : true,
+  "list lattice laws" : true,
+  "map lattice laws" : true,
+  "math lattice laws" : true,
+  "membership of large integral numbers" : true,
+  "number lattice laws" : true,
+  "or" : true,
+  "order" : true,
+  "prune a set by a sign" : true,
+  "prune by a finite set" : true,
+  "prune by an integral domain" : true,
+  "record lattice laws" : true,
+  "signed zeros" : true,
+  "value lattice laws" : true
+}
