--- tests/result/ty/NumericTinyTest
@@ -1,1 +1,0 @@
-ty.NumericTinyTest: 15 / 15
