--- tests/result/ty/NumericTinyTest.json
@@ -1,17 +1,0 @@
-{
-  "NaN" : true,
-  "arithmetic laws" : true,
-  "canonicalize" : true,
-  "exponentiation" : true,
-  "int lattice laws" : true,
-  "join across representations" : true,
-  "math lattice laws" : true,
-  "membership of large integral numbers" : true,
-  "modulo" : true,
-  "number arithmetic laws" : true,
-  "number lattice laws" : true,
-  "prune a set by a sign" : true,
-  "prune by a finite set" : true,
-  "prune by an integral domain" : true,
-  "signed zeros" : true
-}
