--- tests/result/ty/OpTinyTest
@@ -1,1 +1,0 @@
-ty.OpTinyTest: 3 / 3
