--- tests/result/ty/OpTinyTest.json
@@ -1,5 +1,0 @@
-{
-  "and" : true,
-  "or" : true,
-  "order" : true
-}
